fix: Add VAE encoding and patchification for Flux ControlNet control images
CRITICAL FIX: InstantX ControlNet expects VAE-encoded and patchified control
images, not raw RGB. This was causing severe corruption in generated images.

Changes:
1. Added _pack_latents() function for Flux-style 2x2 patchification
   - Converts (B,C,H,W) → (B, (H//2)*(W//2), C*4)
   - Matches diffusers' FluxPipeline implementation

2. Updated compute_control_samples() to accept VAE parameter
   - VAE encodes control image to latent space (16 channels)
   - Applies Flux VAE shift (0.1159) and scaling (0.3611)
   - Patchifies latent to match hidden_states format
   - Falls back to RGB if no VAE (with warning)

3. Access Forge's VAE from sd_model.forge_objects.vae
   - Passed to compute_control_samples()
   - Enables proper control image preprocessing

4. Fixed overlay_canny_edges() int() bug
   - Removed int() call that failed on array indexing
   - Now casts result to uint8 after computation

Why this was needed:
- InstantX ControlNet has no input_hint_block
- Expects control in same format as hidden_states
- Previous implementation passed raw RGB → shape mismatch
- Control samples in wrong scale → image corruption

Reference: https://huggingface.co/InstantX/FLUX.1-dev-Controlnet-Canny/blob/main/config.json

+++ scripts/deforum_helpers/flux_controlnet_preprocessors.py
@@ -165,8 +165,8 @@ def overlay_canny_edges(
     for c in range(3):
         overlay[:, :, c] = np.where(
             edge_mask > 127,  # Edge pixels (white)
-            int(edge_color[c] * alpha + base_image[:, :, c] * (1 - alpha)),
+            edge_color[c] * alpha + base_image[:, :, c] * (1 - alpha),
             base_image[:, :, c]
-        )
+        ).astype(np.uint8)
 
     return overlay
